feat: adding 5 languagess

+++ pkg/ordinal.converter.go
@@ -2,12 +2,12 @@ package pkg
 
 // ToOrdinal converts a number to its ordinal representation
 // Example: 123 -> "123rd", 1 -> "1st", 2 -> "2nd"
-func (op *NumI18NOptions) ToOrdinal(amount float64) string {
+func (op *NumI18NOptions) ToOrdinal(amount int32) string {
 	return ""
 }
 
 // ToOrdinalWords converts a number to its ordinal word representation
 // Example: 123 -> "one hundred twenty-third", 1 -> "first", 2 -> "second"
-func (op *NumI18NOptions) ToOrdinalWords(amount float64) string {
+func (op *NumI18NOptions) ToOrdinalWords(amount int32) string {
 	return ""
 }
